user authentication

+++ base/views.py
@@ -1,9 +1,68 @@
 from django.shortcuts import render, redirect
+from django.contrib import messages
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, Topic
 from .forms import RoomForm
 
 # Create your views here.
+
+
+def loginPage(request):
+    page = 'login'
+    # 이미 로그인한 유저 재로그인 방지
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        # 존재하는 유저인지 확인
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, '존재하지 않는 유저명입니다.')
+
+        # 유저가 존재하면 맞는지 확인하기
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, '잘못된 유저명, 혹은 비밀번호입니다.')
+    context = {
+        'page': page
+    }
+    return render(request, 'base/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            # 데이터 전처리
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, '회원 가입 도중 오류가 발생했습니다. 다시 시도해주세요.')
+
+    return render(request, 'base/login_register.html', {'form': form})
 
 
 
@@ -38,6 +97,8 @@ def room(request, pk):
     return render(request, 'base/room.html', context)
 
 
+# 로그인이 되어있지 않으면 로그인 url로 보내기
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -53,10 +114,15 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(pk=pk)
     # 업데이트할 룸을 지정해주기
     form = RoomForm(instance=room)
+
+    # 유저 확인해주기
+    if request.user != room.host:
+        return HttpResponse('자신이 개설한 방만 편집할 수 있습니다.')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -70,9 +136,16 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(pk=pk)
+    
+    # 유저 확인해주기
+    if request.user != room.host:
+        return HttpResponse('자신이 개설한 방만 편집할 수 있습니다.')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
+
     return render(request, 'base/delete.html', { 'obj' : room })
